+++ resources/js/responsive.js
@@ -1,5 +1,7 @@
 import ResponsiveFieldtype from "./ResponsiveFieldtype";
 import ResponsiveFieldtypeIndex from "./ResponsiveFieldtypeIndex";
 
-Statamic.$components.register('responsive-fieldtype', ResponsiveFieldtype);
-Statamic.$components.register('responsive-fieldtype-index', ResponsiveFieldtypeIndex);
+Statamic.booting(() => {
+    Statamic.$components.register('responsive-fieldtype', ResponsiveFieldtype);
+    Statamic.$components.register('responsive-fieldtype-index', ResponsiveFieldtypeIndex);
+})